added functionality to util

+++ bigmart_sales/R/util.R
@@ -1,5 +1,9 @@
 save_to_csv <- function(df, path) {
   # Add logging
-  write.csv(df, path)
-  message('Adding: ', df)
+  if (!file.exists(path)){
+    write.csv(df, path)
+    message('Adding: ', df)
+  } else {
+    message('File already exists')
+  }
 }
